requirements.txt
tavily-python
langchainhub
python-dotenv
langchain
langchain-openai
langchain-community